Refactor import paths for CLI commands and HTTP modules
