+++ packages/adk-cli/src/http/http.module.ts
@@ -1,14 +1,14 @@
 import { type DynamicModule, Module } from "@nestjs/common";
-import { ConfigModule } from "./modules/config/config.module";
-import { DiscoveryModule } from "./modules/discovery/discovery.module";
-import { EventsModule } from "./modules/events/events.module";
-import { HealthModule } from "./modules/health/health.module";
-import { MessagingModule } from "./modules/messaging/messaging.module";
-import { ProvidersModule } from "./modules/providers/providers.module";
-import { ReloadModule } from "./modules/reload/reload.module";
-import { SessionsModule } from "./modules/sessions/sessions.module";
-import { StateModule } from "./modules/state/state.module";
+import { ConfigModule } from "./config/config.module";
+import { DiscoveryModule } from "./discovery/discovery.module";
+import { EventsModule } from "./events/events.module";
+import { HealthModule } from "./health/health.module";
+import { MessagingModule } from "./messaging/messaging.module";
+import { ProvidersModule } from "./providers/providers.module";
+import { ReloadModule } from "./reload/reload.module";
 import type { RuntimeConfig } from "./runtime-config";
+import { SessionsModule } from "./sessions/sessions.module";
+import { StateModule } from "./state/state.module";
 
 @Module({})
 export class HttpModule {
