+++ packages/adk-cli/src/http/providers/agent-manager.service.ts
@@ -4,8 +4,8 @@ import { pathToFileURL } from "node:url";
 import type { FullMessage, InMemorySessionService } from "@iqai/adk";
 import { AgentBuilder } from "@iqai/adk";
 import { Injectable } from "@nestjs/common";
-import { Logger } from "../../../common/logger";
-import type { Agent, LoadedAgent } from "../../../common/types";
+import { Logger } from "../../common/logger";
+import type { Agent, LoadedAgent } from "../../common/types";
 import { AgentLoader } from "./agent-loader.service";
 import { AgentScanner } from "./agent-scanner.service";
